Dataset
This module includes data collection and understanding data to study the patterns and trends which helps in prediction and evaluating the results. This has been collected using direct questionnaires from patients of Sylhet Diabetes. Hospital in Sylhet, Bangladesh and approved by a doctor. The dataset has many attributes of 521 patients. This dataset consists of 16 features and one target value. Outcome is the feature we are going to predict ‘0’is negative and ‘1’ is
positive. Features of dataset are described below.
1. Age: Age in years ranging from (20years to 65 years)
2. Gender: Male / Female
3. Polyuria: Yes / No
4. Polydipsia: Yes/ No
5. Sudden weight loss: Yes/ No
6. Weakness: Yes/ No
7. Polyphagia: Yes/ No
8. Genital Thrush: Yes/ No
9. Visual blurring: Yes/ No
10. Itching: Yes/ No
11. Irritability: Yes/No
12. Delayed healing: Yes/ No
13. Partial Paresis: Yes/ No
14. Muscle stiffness: yes/ No
15. Alopecia: Yes/ No
16. Obesity: Yes/ No
17. Class: Positive / Negative